move debug callback into gl-util from bootstrap-gl
